[wue] add ezxml support for WIM index parsing
* ezxml was partially altered with some changes from PRs found at https://github.com/lxfontes/ezxml.
* Also apply some small improvements to msapi_utf8.h.

+++ src/msapi_utf8.h
@@ -55,9 +55,9 @@ extern "C" {
 #define LTEXT(txt) _LTEXT(txt)
 
 #define wchar_to_utf8_no_alloc(wsrc, dest, dest_size) \
-	WideCharToMultiByte(CP_UTF8, 0, wsrc, -1, dest, dest_size, NULL, NULL)
+	WideCharToMultiByte(CP_UTF8, 0, wsrc, -1, dest, (int)(dest_size), NULL, NULL)
 #define utf8_to_wchar_no_alloc(src, wdest, wdest_size) \
-	MultiByteToWideChar(CP_UTF8, 0, src, -1, wdest, wdest_size)
+	MultiByteToWideChar(CP_UTF8, 0, src, -1, wdest, (int)(wdest_size))
 #define Edit_ReplaceSelU(hCtrl, str) ((void)SendMessageLU(hCtrl, EM_REPLACESEL, (WPARAM)FALSE, str))
 #define ComboBox_AddStringU(hCtrl, str) ((int)(DWORD)SendMessageLU(hCtrl, CB_ADDSTRING, (WPARAM)FALSE, str))
 #define ComboBox_InsertStringU(hCtrl, index, str) ((int)(DWORD)SendMessageLU(hCtrl, CB_INSERTSTRING, (WPARAM)index, str))
@@ -83,7 +83,7 @@ extern "C" {
 #define wfree(p) sfree(w ## p)
 
 /*
- * Converts an UTF-16 string to UTF8 (allocate returned string)
+ * Converts an UTF-16 string to UTF8 (allocates returned string)
  * Returns NULL on error
  */
 static __inline char* wchar_to_utf8(const wchar_t* wstr)
@@ -115,7 +115,7 @@ static __inline char* wchar_to_utf8(const wchar_t* wstr)
 }
 
 /*
- * Converts an UTF8 string to UTF-16 (allocate returned string)
+ * Converts an UTF8 string to UTF-16 (allocates returned string)
  * Returns NULL on error
  */
 static __inline wchar_t* utf8_to_wchar(const char* str)
@@ -146,7 +146,7 @@ static __inline wchar_t* utf8_to_wchar(const char* str)
 }
 
 /*
-* Converts an non NUL-terminated UTF-16 string of length len to UTF8 (allocate returned string)
+* Converts a non NUL-terminated UTF-16 string of length wlen to NUL-terminated UTF8 (allocates returned string)
 * Returns NULL on error
 */
 static __inline char* wchar_len_to_utf8(const wchar_t* wstr, int wlen)
